feat: Cronjobs to gracefully start/stop containers

+++ run.sh
@@ -1,3 +1,9 @@
 #!/bin/sh
-ansible-galaxy install -r requirements.yml
+
+if [ "$1" = "--force" ]; then
+  ansible-galaxy install -r requirements.yml --force
+else
+  ansible-galaxy install -r requirements.yml
+fi
+
 ansible-playbook lisbon.yml --ask-vault-password --ask-become-pass --vault-id @prompt
